Split large handwritten Go files
