+++ xtrinode/pkg/gateway/query_activity_load.go
@@ -1,0 +1,10 @@
+package gateway
+
+import "time"
+
+// BackendLoad tracks recently observed query activity for a backend.
+type BackendLoad struct {
+	RunningQueries int       `json:"runningQueries"`
+	QueuedQueries  int       `json:"queuedQueries"`
+	LastUpdate     time.Time `json:"lastUpdate"`
+}
